requirements.txt
brotli == 1.*
fonttools == 4.*
Jinja2 == 3.*
pyyaml == 6.*
